fw-6208, extract non-site logo presentation component
