+++ src/components/SiteLogo/SiteLogoPresentation.js
@@ -1,8 +1,9 @@
-import React, { useRef, useEffect, useState } from 'react'
+import React from 'react'
 import PropTypes from 'prop-types'
 import { useParams } from 'react-router-dom'
 
 // FPCC
+import LogoPresentation from 'components/SiteLogo/LogoPresentation'
 import { useSiteStore } from 'context/SiteContext'
 import { IMAGE, MEDIUM, ORIGINAL, SMALL, THUMBNAIL } from 'common/constants'
 import { getMediaPath } from 'common/utils/mediaHelpers'
@@ -11,8 +12,6 @@ import placeholder from 'images/cover-thumbnail.png'
 function SiteLogoPresentation({ logo, size = SMALL, additionalStyling = '' }) {
   const { site } = useSiteStore()
   const { sitename } = useParams()
-  const [loaded, setLoaded] = useState(false)
-  const imgRef = useRef()
 
   const altText = `${site?.title} Logo`
 
@@ -33,32 +32,16 @@ function SiteLogoPresentation({ logo, size = SMALL, additionalStyling = '' }) {
         size,
       })
     }
+
     return placeholder
   }
 
-  useEffect(() => {
-    if (imgRef?.current?.complete) {
-      setLoaded(true)
-    }
-  }, [])
-
   return (
-    <div
-      data-testid="SiteLogoPresentation"
-      className={`relative overflow-hidden w-full aspect-w-1 aspect-h-1 rounded-full bg-gray-50 ${additionalStyling}`}
-    >
-      <div style={{ paddingBottom: '100%' }} />
-      <img
-        loading="lazy"
-        src={getLogoSrc()}
-        alt={altText}
-        ref={imgRef}
-        onLoad={() => setLoaded(true)}
-        className={`absolute w-full h-full top-0 bottom-0 left-0 right-0 object-cover object-center ${
-          loaded ? 'opacity-1' : 'opacity-0'
-        }`}
-      />
-    </div>
+    <LogoPresentation
+      imgSrc={getLogoSrc()}
+      altText={altText}
+      additionalStyling={additionalStyling}
+    />
   )
 }
